requirements.txt
Flask==0.10.1
Flask-Admin==1.0.7
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.21
SQLAlchemy==0.9.4
WTForms==1.0.5
Werkzeug==0.9.4
ago==0.0.6
alembic==0.6.4
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.24
oslo.config==1.3.0
six==1.6.1
wsgiref==0.1.2
